\name{bivrp-internal}
\alias{add.dplots.plot}
\alias{add.dplots.prep}
\alias{chp.perpoint}
\alias{sorttheta}
\encoding{UTF-8}
\title{Internal functions to prepare \code{bivrp} objects}
\description{Internal functions used to prepare \code{bivrp} objects for plotting}
\author{
Rafael A. Moral <rafael_moral@yahoo.com.br>, John Hinde and Clarice G. B. \enc{Demétrio}{Demetrio}
}