\name{bivrp-package}
\alias{bivrp-package}
\encoding{UTF-8}
\docType{package}
\title{
\packageTitle{bivrp}
}
\description{
\packageDescription{bivrp}
}
\details{

\packageDESCRIPTION{bivrp}
\packageIndices{bivrp}
}
\author{
\packageAuthor{bivrp}

Maintainer: \packageMaintainer{bivrp}
}
\keyword{ package }