\name{bivrp-internal}
\alias{add_dplots_plot}
\alias{add_dplots_prep}
\alias{chp_perpoint}
\alias{sort_theta}
\encoding{UTF-8}
\title{Internal functions to prepare \code{bivrp} objects}
\description{Internal functions used to prepare \code{bivrp} objects for plotting}
\author{
Rafael A. Moral <rafael.deandrademoral@mu.ie> and John Hinde
}